Create a fully statically linked bpftrace binary
We want bpftrace to be completely statically linked, so as to prevent
any binary incompatibilities from creeping in. Unfortunately, that seems
to be impossible with LTO enabled, as we run into a bunch of linker
errors of the form:
> ld.lld: error: bionic/libc/arch-arm64/bionic/setjmp.S:199:(.text+0x29C): relocation R_AARCH64_CONDBR19 out of range: 14118624 is not in [-1048576, 1048575]

It seems as if switching to a different linker may fix the problem, but
that is out of the scope of this work. For now, disable LTO, at the
expense of a slight increase (~2 MiB) of the binary (which now of course
includes parts of libc++, libc, and libm).